+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- database: Fix video rotation
 - database: Fix video duration
 
 ## [1.10.0] - 2023-06-13